backend/requirements.txt
asgiref==3.7.2
astroid==3.0.3
dill==0.3.8
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
isort==5.13.2
mccabe==0.7.0
platformdirs==4.2.0
PyJWT==2.8.0
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-decouple==3.8
pytz==2023.4
sqlparse==0.4.4
tomlkit==0.12.3
tzdata==2023.4
